fix(app): always assign a liquid color if one is not passed in
