+++ app/src/organisms/ODD/ProtocolSetup/ProtocolSetupLabware/LabwareMapView.tsx
@@ -56,7 +56,8 @@ export function LabwareMapView(props: LabwareMapViewProps): JSX.Element {
           ? getWellFillFromLabwareId(
               topLabwareInfo.labwareId,
               mostRecentAnalysis?.liquids ?? [],
-              labwareByLiquidId
+              labwareByLiquidId,
+              mostRecentAnalysis?.commands ?? []
             )
           : undefined
       return {
@@ -94,7 +95,8 @@ export function LabwareMapView(props: LabwareMapViewProps): JSX.Element {
     const wellFill = getWellFillFromLabwareId(
       topLabwareInfo.labwareId,
       mostRecentAnalysis?.liquids ?? [],
-      labwareByLiquidId
+      labwareByLiquidId,
+      mostRecentAnalysis?.commands ?? []
     )
 
     return {
