+++ shared-data/js/helpers/getWellFillFromLabwareId.ts
@@ -1,3 +1,6 @@
+import { parseLiquidsInLoadOrder } from './parseProtocolCommands'
+
+import type { RunTimeCommand } from '../../protocol'
 import type { Liquid } from '../types'
 import type { LabwareByLiquidId } from './getLabwareInfoByLiquidId'
 
@@ -5,10 +8,12 @@ export type WellFill = Record<string, string>
 
 export function getWellFillFromLabwareId(
   labwareId: string,
-  liquidsInLoadOrder: Liquid[],
-  labwareByLiquidId: LabwareByLiquidId
+  liquids: Liquid[],
+  labwareByLiquidId: LabwareByLiquidId,
+  commands: RunTimeCommand[]
 ): WellFill {
   let labwareWellFill: WellFill = {}
+  const liquidsInLoadOrder = parseLiquidsInLoadOrder(liquids, commands)
   const liquidIds = Object.keys(labwareByLiquidId)
   const labwareInfo = Object.values(labwareByLiquidId)
